% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfredx.R
\name{alfredx}
\alias{alfredx}
\title{Retrieve data from FRED}
\usage{
alfredx(series_id, start_date = NULL, end_date = NULL, api_key = NULL)
}
\arguments{
\item{series_id}{FRED series id}

\item{start_date}{A date object}

\item{end_date}{A date object}

\item{api_key}{FRED API key}
}
\value{
A tibble
}
\description{
Retrieve data from FRED
}
